bpo-17005: Move topological sort functionality to its own module
